Refactor scoring parameters structure and update related services
- Introduced NestedScoringParameters and ScoringRule classes to better organize scoring parameters.
- Updated ProjetosService to utilize the new scoring parameters structure.
- Refactored ScoringService to accommodate the new nested scoring parameters.
- Modified API calls and context to handle NestedScoringParameters.
- Updated frontend components to work with the new scoring parameters structure.
- Removed deprecated scoring parameters handling and ensured backward compatibility where necessary.

+++ AvaliaDados/src/main/java/com/avaliadados/model/CollaboratorEntity.java
@@ -29,7 +29,7 @@ public class CollaboratorEntity {
     @CollectionTable(name = "collaborator_parametros", joinColumns = @JoinColumn(name = "collaborator_id"))
     @MapKeyColumn(name = "param_key")
     @Column(name = "param_value")
-    private Map<String, Double> parametros;
+    private Map<String, Integer> parametros;
 
 
     public CollaboratorEntity(String nome, String cpf, String idCallRote, int pontuacao, String role, Long version) {
